Revise build order to RGB to match RGB bit load routine
Changes the order of build to match the order of the bit shift routine.
The previous shift routine "forces" an arrangement to a RGB order based
on color order selected by user but the build actually loads it in GRB
order.  This change will build as the bits were populated.