Fix fread ignored result (#370)
Some compilers (e.g. gcc-7.4.0 on Ubuntu 18.04 aarch64) complain about
the fact that fread()'s result is ignored. This breaks compilation. Fix
this so that the library can be compiled with more compilers.

This issue was found by a user who was accidentally(?) installing this
library on NVIDIA Jetson. I haven't evaluated whether this library is
useful on this platform, but there's no reason it should fail to compile
there for such a generic issue. Besides, this fix will probably be useful
on more relevant platforms if/when they upgrade their compilers.